requirements.txt
beautifulsoup4==4.9.3
Brotli==1.0.9
certifi==2020.12.5
chardet==4.0.0
idna==2.10
python-magic-bin==0.4.14
requests==2.25.1
soupsieve==2.2.1
urllib3==1.26.4
